requirements.txt
streamlit
tensorflow
matplotlib
seaborn
scikit-learn
Pillow
numpy
requests
